+++ com.unity.render-pipelines.core/Editor/ShaderGenerator/ShaderGeneratorMenu.cs
@@ -5,7 +5,7 @@ namespace UnityEditor.Rendering
 {
     class ShaderGeneratorMenu
     {
-        [MenuItem("Edit/Render Pipeline/Generate Shader Includes", priority = CoreUtils.editMenuPriority1)]
+        [MenuItem("Edit/Rendering/Generate Shader Includes", priority = CoreUtils.Sections.k_Section3 + CoreUtils.Priorities.k_EditMenuPriority + 1)]
         async static Task GenerateShaderIncludes()
         {
             await CSharpToHLSL.GenerateAll();